# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RelaySMS Publisher",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/stats/publications": {
      "get": {
        "tags": [
          "Stats"
        ],
        "summary": "List publications",
        "description": "Newest first. Scope: stats:publications:read.\n\n`failure_reason` is null without stats:publications:reasons.",
        "operationId": "list_publication_stats_v1_stats_publications_get",
        "security": [
          {
            "HTTPBasic": [
              "stats:publications:read"
            ]
          },
          {
            "SessionCookie": [
              "stats:publications:read"
            ]
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "description": "Set by the next and prev links",
              "title": "Cursor"
            },
            "description": "Set by the next and prev links"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20,
                  "pattern": "^[a-zA-Z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status",
              "title": "Status"
            },
            "description": "Filter by status"
          },
          {
            "name": "platform_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100,
                  "pattern": "^[a-zA-Z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by platform name",
              "title": "Platform Name"
            },
            "description": "Filter by platform name"
          },
          {
            "name": "protocol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20,
                  "pattern": "^[a-zA-Z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by ingestion protocol",
              "title": "Protocol"
            },
            "description": "Filter by ingestion protocol"
          },
          {
            "name": "country_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 10,
                  "pattern": "^[a-zA-Z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by ISO country code",
              "title": "Country Code"
            },
            "description": "Filter by ISO country code"
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Created at or after (ISO-8601, UTC if no offset)",
              "title": "Since"
            },
            "description": "Created at or after (ISO-8601, UTC if no offset)"
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Created before (ISO-8601, UTC if no offset)",
              "title": "Until"
            },
            "description": "Created before (ISO-8601, UTC if no offset)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationStatsPage"
                }
              }
            }
          },
          "400": {
            "description": "An invalid cursor, or 'since' not before 'until'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks stats:publications:read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stats/publications/summary": {
      "get": {
        "tags": [
          "Stats"
        ],
        "summary": "Summarize publications",
        "description": "Counts per group; last 30 days by default. Scope: stats:publications:read.\n\nGrouping by `failure_reason` also needs stats:publications:reasons. Groups sort by\nperiod, then count, highest first. Periods with no rows are left out; chart them\nas 0.",
        "operationId": "summarize_publication_stats_v1_stats_publications_summary_get",
        "security": [
          {
            "HTTPBasic": [
              "stats:publications:read"
            ]
          },
          {
            "SessionCookie": [
              "stats:publications:read"
            ]
          }
        ],
        "parameters": [
          {
            "name": "group_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/StatsGroupBy"
              },
              "description": "Columns to group by. Repeatable.",
              "default": [
                "status"
              ],
              "title": "Group By"
            },
            "description": "Columns to group by. Repeatable."
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatsInterval"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Also group by period start, in UTC. Weeks start Monday.",
              "title": "Interval"
            },
            "description": "Also group by period start, in UTC. Weeks start Monday."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20,
                  "pattern": "^[a-zA-Z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status",
              "title": "Status"
            },
            "description": "Filter by status"
          },
          {
            "name": "platform_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100,
                  "pattern": "^[a-zA-Z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by platform name",
              "title": "Platform Name"
            },
            "description": "Filter by platform name"
          },
          {
            "name": "protocol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20,
                  "pattern": "^[a-zA-Z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by ingestion protocol",
              "title": "Protocol"
            },
            "description": "Filter by ingestion protocol"
          },
          {
            "name": "country_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 10,
                  "pattern": "^[a-zA-Z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by ISO country code",
              "title": "Country Code"
            },
            "description": "Filter by ISO country code"
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Created at or after (ISO-8601, UTC if no offset)",
              "title": "Since"
            },
            "description": "Created at or after (ISO-8601, UTC if no offset)"
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Created before (ISO-8601, UTC if no offset)",
              "title": "Until"
            },
            "description": "Created before (ISO-8601, UTC if no offset)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationStatsSummary"
                }
              }
            }
          },
          "400": {
            "description": "'since' isn't before 'until'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks stats:publications:read. Also when grouping by failure_reason without stats:publications:reasons.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/publications": {
      "post": {
        "tags": [
          "Publishing"
        ],
        "summary": "Publish",
        "description": "Queues an SMS payload for publishing.\n\nMulti-part payloads are kept until every segment arrives, then published. Payloads\nare tagged with protocol `https`; see Offline Publishing in the README for\n`OFFLINE_PUBLISH_ALLOWED_PROTOCOLS` and `tag`. Decryption and adapter errors\nhappen later in the queue, so they're logged, not returned.",
        "operationId": "create_publications_v1_publications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRestContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishContentResponse"
                }
              }
            }
          },
          "400": {
            "description": "The payload isn't base64, or isn't a RelaySMS payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "What went wrong, safe to show to users."
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The body of every error response."
      },
...
      "PublicationStat": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "platform_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Name"
          },
          "protocol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at"
        ],
        "title": "PublicationStat"
      },
      "PublicationStatsPage": {
        "properties": {
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next",
            "description": "URL of the next (older) page; null on the last page."
          },
          "prev": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev",
            "description": "URL of the previous (newer) page; null on the first page."
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/PublicationStat"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "PublicationStatsPage"
      },
      "PublicationStatsSummary": {
        "properties": {
          "since": {
            "type": "string",
            "format": "date-time",
            "title": "Since"
          },
          "until": {
            "type": "string",
            "format": "date-time",
            "title": "Until"
          },
          "interval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatsInterval"
              },
              {
                "type": "null"
              }
            ]
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/PublicationStatsSummaryGroup"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "type": "object",
        "required": [
          "since",
          "until",
          "total",
          "groups"
        ],
        "title": "PublicationStatsSummary"
      },
      "PublicationStatsSummaryGroup": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "count"
        ],
        "title": "PublicationStatsSummaryGroup",
        "description": "group_by column values plus their count."
      },
      "PublishContentResponse": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Always null here; errors come back with a 4xx status."
          }
        },
        "type": "object",
        "title": "PublishContentResponse"
      },
      "PublishRestContentRequest": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address",
            "description": "Sender phone number in E.164 format",
            "examples": [
              "+12025550123"
            ]
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Base64-encoded SMS payload"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag",
            "description": "Shared secret for offline payloads, if the server sets one."
          }
        },
        "type": "object",
        "required": [
          "address",
          "text"
        ],
        "title": "PublishRestContentRequest"
      },
...
      "StatsGroupBy": {
        "type": "string",
        "enum": [
          "status",
          "platform_name",
          "protocol",
          "country_code",
          "failure_reason"
        ],
        "title": "StatsGroupBy"
      },
      "StatsInterval": {
        "type": "string",
        "enum": [
          "day",
          "week",
          "month",
          "year"
        ],
        "title": "StatsInterval"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "SessionCookie": {
        "type": "apiKey",
        "description": "Set by POST /v1/auth/login.",
        "in": "cookie",
        "name": "relaysms_session"
      }
    }
  }
}